# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillMatch AI Services API",
    "description": "Production AI Services API for SkillMatch. The canonical AI-Serv5 CV extraction pipeline powers CV, matching, interview, and review features.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/recommendations/feed": {
      "get": {
        "tags": [
          "Personalized Job Recommendations"
        ],
        "summary": "Get Personalized Job Recommendation Feed",
        "description": "Returns an ordered list of recommended jobs for a candidate. Scores jobs deterministically using candidate skills, target role, preferences, match quality, freshness, and interaction history. Excludes applied/dismissed jobs and deduplicates identical postings.",
        "operationId": "get_recommendation_feed_api_v1_recommendations_feed_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Unique identifier of the candidate",
              "title": "Candidate Id"
            },
            "description": "Unique identifier of the candidate"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Number of recommendations per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of recommendations per page"
          },
          {
            "name": "work_mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional work mode filter override (remote, hybrid, onsite)",
              "title": "Work Mode"
            },
            "description": "Optional work mode filter override (remote, hybrid, onsite)"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional location filter substring",
              "title": "Location"
            },
            "description": "Optional location filter substring"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 100.0,
              "minimum": 0.0,
              "description": "Minimum recommendation score cutoff",
              "default": 0.0,
              "title": "Min Score"
            },
            "description": "Minimum recommendation score cutoff"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationFeedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Candidate not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/events": {
      "post": {
        "tags": [
          "Personalized Job Recommendations"
        ],
        "summary": "Record Candidate Job Interaction Event",
        "description": "Internal service endpoint for recording a candidate interaction with a job (view, click, save, unsave, apply, dismiss, undismiss). The trusted caller must authenticate and authorize the acting candidate before forwarding this request. Validates that both the candidate and job exist in persistence, enforces consistent state transitions, and updates the candidate's interaction history used by recommendation scoring.",
        "operationId": "record_recommendation_event_api_v1_recommendations_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Interaction recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsupported event type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing service API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Candidate or Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error in request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service API key is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExtractionErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Description of the error that occurred"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Standardized error code",
            "example": "UNSUPPORTED_FILE_TYPE"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error_code"
        ],
        "title": "ExtractionErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationEventRequest": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "minLength": 1,
            "title": "Candidate Id",
            "description": "Unique identifier of the candidate"
          },
          "job_id": {
            "type": "string",
            "minLength": 1,
            "title": "Job Id",
            "description": "Unique identifier of the job"
          },
          "event_type": {
            "type": "string",
            "minLength": 1,
            "title": "Event Type",
            "description": "Interaction type: view, click, save, unsave, apply, dismiss, or undismiss; accepted aliases are normalized."
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Optional interaction metadata"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "job_id",
          "event_type"
        ],
        "title": "RecommendationEventRequest",
        "description": "Payload for submitting a candidate interaction event."
      },
      "RecommendationEventResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the event was successfully recorded",
            "default": true
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id",
            "description": "Identifier of the candidate"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Identifier of the job"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "description": "Recorded canonical interaction type"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At",
            "description": "Timestamp when the event was recorded"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message",
            "default": "Interaction recorded successfully."
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "job_id",
          "event_type",
          "recorded_at"
        ],
        "title": "RecommendationEventResponse",
        "description": "Response returned upon recording a candidate interaction event."
      },
      "RecommendationExplanation": {
        "properties": {
          "headline": {
            "type": "string",
            "title": "Headline",
            "description": "Concise summary headline of why this job is recommended"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/RecommendationReason"
            },
            "type": "array",
            "title": "Reasons",
            "description": "Structured factual reasons"
          },
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Skills",
            "description": "Names of candidate skills matching job requirements"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills",
            "description": "Names of required job skills missing from candidate profile"
          }
        },
        "type": "object",
        "required": [
          "headline"
        ],
        "title": "RecommendationExplanation",
        "description": "Factual explanation containing structured reasons and concrete matched/missing skills."
      },
      "RecommendationFeedResponse": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id",
            "description": "Unique candidate identifier"
          },
          "total_results": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Results",
            "description": "Total matching jobs available after filtering"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "Current page number",
            "default": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum items requested per page",
            "default": 20
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether more pages are available",
            "default": false
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At",
            "description": "UTC timestamp of feed generation"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendedJobItem"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Ordered list of recommended job items (DAY 1 root field name)"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "total_results"
        ],
        "title": "RecommendationFeedResponse",
        "description": "Top-level recommendation feed response payload."
      },
      "RecommendationReason": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable reason code, e.g. SKILL_MATCH_HIGH, TARGET_ROLE_MATCH"
          },
          "category": {
            "type": "string",
            "enum": [
              "skills",
              "role",
              "experience",
              "preference",
              "freshness",
              "behavior"
            ],
            "title": "Category",
            "description": "Category of the recommendation signal"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Human-readable explanation text"
          },
          "weight": {
            "type": "string",
            "enum": [
              "primary",
              "secondary",
              "supporting"
            ],
            "title": "Weight",
            "description": "Importance weight of this specific reason",
            "default": "supporting"
          }
        },
        "type": "object",
        "required": [
          "code",
          "category",
          "text"
        ],
        "title": "RecommendationReason",
        "description": "Structured evidence-based reason for recommending a job."
      },
      "RecommendedJobItem": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Unique job identifier"
          },
          "rank": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rank",
            "description": "1-indexed rank position in the feed"
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "description": "Composite deterministic recommendation score (0-100)"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons",
            "description": "Concise string list of recommendation reasons (DAY 1 compatible)"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Job title",
            "default": ""
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company",
            "description": "Hiring company name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Geographic location / city"
          },
          "work_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Mode",
            "description": "remote, hybrid, or onsite"
          },
          "employment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employment Type",
            "description": "full_time, part_time, contract, internship"
          },
          "experience_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level",
            "description": "Entry, Mid, Senior, Lead"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At",
            "description": "Timestamp when the job was posted"
          },
          "qualification_status": {
            "type": "string",
            "title": "Qualification Status",
            "description": "Verdict from MatchingService: Qualified, Partially Qualified, Not Qualified",
            "default": "Qualified"
          },
          "score_breakdown": {
            "$ref": "#/components/schemas/ScoreBreakdown",
            "description": "Granular component scores"
          },
          "explanation": {
            "$ref": "#/components/schemas/RecommendationExplanation",
            "description": "Evidence-based explanation breakdown"
          },
          "is_saved": {
            "type": "boolean",
            "title": "Is Saved",
            "description": "Whether the candidate previously saved this job",
            "default": false
          },
          "is_applied": {
            "type": "boolean",
            "title": "Is Applied",
            "description": "Whether the candidate has applied to this job",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "rank",
          "score",
          "score_breakdown",
          "explanation"
        ],
        "title": "RecommendedJobItem",
        "description": "Individual recommended job entry in the feed."
      },
...
      "ScoreBreakdown": {
        "properties": {
          "match_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Match Score",
            "description": "Core technical qualification score from MatchingService (0-100)"
          },
          "role_relevance_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Role Relevance Score",
            "description": "Target role alignment score (0-100)"
          },
          "preference_fit_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Preference Fit Score",
            "description": "Work mode, location, and employment type fit (0-100)"
          },
          "freshness_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Freshness Score",
            "description": "Exponential decay recency score (0-100)"
          },
          "behavior_boost": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Behavior Boost",
            "description": "Interaction affinity score (0-100, neutral=50.0)",
            "default": 50.0
          }
        },
        "type": "object",
        "required": [
          "match_score",
          "role_relevance_score",
          "preference_fit_score",
          "freshness_score"
        ],
        "title": "ScoreBreakdown",
        "description": "Normalized sub-scores (0.0 - 100.0) contributing to the final recommendation score."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}